Fix error in which book showed Mom I read quickly
I wasn't sure whether it was Ben-Hur or 20,000 Leagues when I first made
this page, so I asked her. She thought 20,000 Leagues.

On further reflection, I remembered that it actually took me more than
two days to read 20,000 Leagues, and that she was confusing this event
with a different anecdote (the one where she asked me if Captain Nemo
was a bad guy or not, and I said "I'm not sure," with the book mostly
done, reflecting a grasp of the book's essential conflict [which is
Nemo's morality or lack thereof]).